backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
djangorestframework-simplejwt>=5.3
celery>=5.4
redis>=5.0
gunicorn>=22.0
python-dotenv>=1.0
sentry-sdk[django]>=2.0